refactor(api): split sync.ts into sub-routers; governance gets a real home (A17.6)
The 1713-line api/sync.ts with 24 routes is now api/sync/:
  docs.ts (13)       — the four record families + batch-upsert
  tombstones.ts (4)  — record and file tombstones
  manifest.ts (2)    — file manifest + merkle summary
  members.ts (2), votes.ts (2), warm.ts (1)
  _shared.ts         — incoming-doc schemas, peer/space authorisation, cursor
                       codec, fork-depth + implausible-seq guards, link-violation
                       recorders

Every URL is unchanged. The ejection guard — a router-level `syncRouter.use(...)`
middleware that 401s any request scoped to a network this instance was ejected
from — stays on the parent in index.ts, still registered ahead of every
sub-router. A route-only extractor drops that middleware silently and still
compiles; the completeness assertion added in A17.5 is what caught it.

concludeRoundIfReady and sendMemberRemovedNotify now live in sync/governance.ts.
They were exported from a ROUTE module and imported by five others (api/invite.ts,
api/networks/{join,members,votes}.ts, sync/engine.ts). Round state is config
(network.pendingRounds), not in-memory, so both the HTTP surfaces and the engine
can drive conclusion — the helpers just needed a home. This resolves the coupling
A17.5 deliberately left.

Also: router names are now asserted UNIQUE across server/src/api. The audit and
guard analyses key prefixes by router name, so duplicates silently give one router
the other's prefix. That bit twice: A17.3 (filesRouter x2) and A17.6
(membersRouter/votesRouter in both api/networks/ and api/sync/, which made the peer
routes report as unaudited). Both compiled fine. The sync sub-routers are now
sync*Router, the /api/sync peer-auth exemption follows them, and a new assertion
fails on any duplicate — verified by reintroducing the collision and watching it
fail, then pass once fixed.

Docs: sync-protocol.md claimed sendMemberRemovedNotify was "called from three
places: ... sync.ts ... networks.ts ... engine.ts". Those files no longer exist and
there are in fact four call sites. Corrected, with its new module home.

Verified: build clean, no dead code (tsc --noUnusedLocals); route parity 24/24;
sync 191/0 (1 skipped), red-team 272/272, standalone 801/0 (805 total — +1 for the
new uniqueness guard), brain 183/183, mcp-tools 94/94, proxy-spaces 44/44,
schema 29+85, conflict 15, export 14, conversion 8.
