+++ server/src/app.ts
@@ -9,7 +9,7 @@ import { spacesRouter } from './api/spaces.js';
 import { fileStoreRouter } from './api/files.js';
 import { conflictsRouter } from './api/conflicts.js';
 import { duplicatesRouter } from './api/duplicates.js';
-import { syncRouter } from './api/sync.js';
+import { syncRouter } from './api/sync/index.js';
 import { networksRouter } from './api/networks/index.js';
 import { notifyRouter } from './api/notify.js';
 import { inviteRouter } from './api/invite.js';
